refactor: clean auth-service structure
