+++ authentication/src/main/java/com/learnandphish/authentication/jwt/JwtResponse.java
@@ -1,4 +1,4 @@
-package com.learnandphish.authentication;
+package com.learnandphish.authentication.jwt;
 
 public record JwtResponse(String jwtToken) {
