+++ authentication/src/main/java/com/learnandphish/authentication/jwt/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.learnandphish.authentication;
+package com.learnandphish.authentication.jwt;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
