+++ authentication/src/main/java/com/learnandphish/authentication/user/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.learnandphish.authentication;
+package com.learnandphish.authentication.user;
 
 public class RegisterRequest {
