+++ authentication/src/main/java/com/learnandphish/authentication/user/Roles.java
@@ -1,4 +1,4 @@
-package com.learnandphish.authentication;
+package com.learnandphish.authentication.user;
 
 /**
  *
