+++ authentication/src/test/java/com/learnandphish/authentication/JwtUserDetailsServiceTest.java
@@ -1,8 +1,11 @@
 package com.learnandphish.authentication;
 
+import com.learnandphish.authentication.jwt.JwtUserDetailsService;
+import com.learnandphish.authentication.user.Roles;
+import com.learnandphish.authentication.user.UserData;
+import com.learnandphish.authentication.user.UserDataRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
